+++ CHEWBBACA/SchemaEvaluator/resources/schema_evaluator_individual_reports_react_scripts/webpack.config.js
@@ -12,7 +12,7 @@ module.exports = {
   },
   module: {
     rules: [
-      { test: /\.js$/, exclude: /node_modules/, use: "babel-loader" },
+      { test: /\.(js|jsx)$/, exclude: /node_modules/, use: "babel-loader" },
       {
         test: /\.scss$/,
         use: ["style-loader", "css-loader", "sass-loader"],