+++ CHEWBBACA/chewBBACA.py
@@ -180,6 +180,7 @@ def create_schema():
         os.makedirs(args.output_directory)
 
     genomes_list = os.path.join(args.output_directory, 'listGenomes2Call.txt')
+    input_type = os.path.isdir(args.input_files)
     args.input_files = pv.check_input_type(args.input_files, genomes_list)
 
     # add clustering config
@@ -213,7 +214,8 @@ def create_schema():
 
     # remove temporary file with paths
     # to genome files
-    os.remove(args.input_files)
+    if input_type is True:
+        os.remove(args.input_files)
 
 
 @pd.process_timer
@@ -415,9 +417,12 @@ def allele_call():
     # if not check if is a folder or a txt with a list of paths
     if args.genes_list is not False:
         schema_genes = pv.check_input_type(args.genes_list, 'listGenes2Call.txt', args.schema_directory)
+        input_type = [False]
     else:
         schema_genes = pv.check_input_type(args.schema_directory, 'listGenes2Call.txt')
+        input_type = [True]
 
+    input_type.append(os.path.isdir(args.input_files))
     genomes_files = pv.check_input_type(args.input_files, 'listGenomes2Call.txt')
 
     # determine if schema was downloaded from Chewie-NS
@@ -442,7 +447,10 @@ def allele_call():
     #     updated = ps.store_allelecall_results(args.output_directory, args.schema_directory)
 
     # remove temporary files with paths to genomes and schema files
-    fo.remove_files([schema_genes, genomes_files])
+    if input_type[0] is True:
+        fo.remove_files([schema_genes])
+    if input_type[1] is True:
+        fo.remove_files([genomes_files])
 
 
 @pd.process_timer
@@ -799,8 +807,10 @@ def remove_genes():
 
         return usage_msg
 
-    parser = argparse.ArgumentParser(description='Remove loci from a matrix with allelic profiles.',
-                                     usage=msg())
+    parser = argparse.ArgumentParser(prog='RemoveGenes',
+                                     description='Remove loci from a matrix with allelic profiles.',
+                                     usage=msg(),
+                                     formatter_class=ModifiedHelpFormatter)
 
     parser.add_argument('RemoveGenes', nargs='+',
                         help='Remove loci from a matrix with allelic profiles.')
@@ -812,12 +822,12 @@ def remove_genes():
 
     parser.add_argument('-g', '--genes-list', type=str,
                         required=True, dest='genes_list',
-                        help='File with the list of genes to remove.')
+                        help='File with the list of genes to remove, one identifier '
+                             'per line.')
 
     parser.add_argument('-o', '--output-file', type=str,
                         required=True, dest='output_file',
-                        help='Path to the output file that will be created with the '
-                             'new matrix.')
+                        help='Path to the output file.')
 
     parser.add_argument('--inverse', action='store_true',
                         default=False, dest='inverse',
@@ -834,32 +844,37 @@ def remove_genes():
 def join_profiles():
 
     def msg(name=None):
-        return '''chewBBACA.py JoinProfiles [RemoveGenes ...][-h]
-                  -p1 -p2 -o [O]'''
 
-    parser = argparse.ArgumentParser(description='This program joins two '
-                                                 'profiles, returning a '
-                                                 'single profile file with '
-                                                 'the common loci',
-                                     usage=msg())
+        # simple command to adapt external schema with default arguments values
+        simple_cmd = ('  chewBBACA.py JoinProfiles -p <profiles1> <profiles2> '
+                      '-o <output_file> ')
+
+        usage_msg = ('\nJoin allele calling results from two runs:\n\n{0}\n'.format(simple_cmd))
+
+        return usage_msg
+
+    parser = argparse.ArgumentParser(prog='JoinProfiles',
+                                     description='Joins allele calling results from '
+                                                 'multiple runs.',
+                                     usage=msg(),
+                                     formatter_class=ModifiedHelpFormatter)
 
     parser.add_argument('JoinProfiles', nargs='+',
-                        help='join profiles')
+                        help='')
 
-    parser.add_argument('-p1', '--profile1', type=str,
-                        required=True, dest='profile1',
-                        help='Path to file containing a matrix '
-                             'with allelic profiles.')
-
-    parser.add_argument('-p2', '--profile2', type=str,
-                        required=True, dest='profile2',
-                        help='Path to file containing a matrix '
-                             'with allelic profiles.')
+    parser.add_argument('-p', '--profiles', nargs='+', type=str,
+                        required=True, dest='profiles',
+                        help='Path to files containing the results from '
+                             'the AlleleCall process (results_alleles.tsv).')
 
     parser.add_argument('-o', '--output-file', type=str,
                         required=True, dest='output_file',
-                        help='Paht to output file with result '
-                             'of joining both inputs.')
+                        help='Path to the output file.')
+
+    parser.add_argument('--common', action='store_true',
+                        required=False, dest='common',
+                        help='Create file with profiles for '
+                             'the set of common loci.')
 
     args = parser.parse_args()
     del args.JoinProfiles
@@ -1456,8 +1471,8 @@ def main():
                       'PrepExternalSchema': ['Adapt an external schema to be '
                                              'used with chewBBACA.',
                                              prep_schema],
-                      'JoinProfiles': ['Join two profiles in a single profile '
-                                       'file.',
+                      'JoinProfiles': ['Joins allele calling results from '
+                                       'different runs.',
                                        join_profiles],
                       'UniprotFinder': ['Retrieve annotations for loci in a schema.',
                                         find_uniprot],